Rename gameActions to game
